Add namespace attribute to cancel/timeout child spans
Address review: the CustomRun/TaskRun child spans in cancelCustomRun,
cancelTaskRun, timeoutCustomRun and timeoutTaskRun only recorded the run
name. Add the namespace attribute (already passed in) for consistency with
the other reconciler spans and so the spans are filterable on their own.

Also move TestTimeoutPipelineRunEmitsSpan to timeout_test.go next to the
other timeout tests; the shared spanNames helper stays in cancel_test.go.
