+++ pkg/reconciler/pipelinerun/cancel_test.go
@@ -458,38 +458,6 @@ func TestCancelPipelineRunEmitsSpan(t *testing.T) {
 	}
 }
 
-func TestTimeoutPipelineRunEmitsSpan(t *testing.T) {
-	sr := tracetest.NewSpanRecorder()
-	tp := tracesdk.NewTracerProvider(tracesdk.WithSpanProcessor(sr))
-
-	pr := &v1.PipelineRun{
-		ObjectMeta: metav1.ObjectMeta{Name: "test-pipeline-run-timeout", Namespace: "foo"},
-	}
-
-	ctx, _ := ttesting.SetupFakeContext(t)
-	cfg := config.NewStore(logtesting.TestLogger(t))
-	ctx = cfg.ToContext(ctx)
-	// Put a span on the context so the timeout helpers create their child spans
-	// against the recording provider.
-	ctx, root := tp.Tracer("test").Start(ctx, "root")
-	defer root.End()
-	c, _ := test.SeedTestData(t, ctx, test.Data{PipelineRuns: []*v1.PipelineRun{pr}})
-
-	if err := timeoutPipelineRun(ctx, logtesting.TestLogger(t), pr, c.Pipeline); err != nil {
-		t.Fatalf("timeoutPipelineRun returned an unexpected error: %v", err)
-	}
-
-	var found bool
-	for _, s := range sr.Ended() {
-		if s.Name() == "timeoutPipelineRun" {
-			found = true
-		}
-	}
-	if !found {
-		t.Errorf("expected a %q span to be recorded, got %v", "timeoutPipelineRun", spanNames(sr.Ended()))
-	}
-}
-
 func spanNames(spans []tracesdk.ReadOnlySpan) []string {
 	names := make([]string, 0, len(spans))
 	for _, s := range spans {
