+++ pkg/reconciler/pipelinerun/timeout_test.go
@@ -17,12 +17,15 @@ import (
 	"context"
 	"testing"
 
+	"github.com/tektoncd/pipeline/pkg/apis/config"
 	"github.com/tektoncd/pipeline/pkg/apis/pipeline"
 	v1 "github.com/tektoncd/pipeline/pkg/apis/pipeline/v1"
 	"github.com/tektoncd/pipeline/pkg/apis/pipeline/v1beta1"
 	_ "github.com/tektoncd/pipeline/pkg/pipelinerunmetrics/fake" // Make sure the pipelinerunmetrics are setup
 	ttesting "github.com/tektoncd/pipeline/pkg/reconciler/testing"
 	"github.com/tektoncd/pipeline/test"
+	tracesdk "go.opentelemetry.io/otel/sdk/trace"
+	"go.opentelemetry.io/otel/sdk/trace/tracetest"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"knative.dev/pkg/apis"
@@ -300,5 +303,37 @@ func TestTimeoutPipelineRun(t *testing.T) {
 				}
 			}
 		})
+	}
+}
+
+func TestTimeoutPipelineRunEmitsSpan(t *testing.T) {
+	sr := tracetest.NewSpanRecorder()
+	tp := tracesdk.NewTracerProvider(tracesdk.WithSpanProcessor(sr))
+
+	pr := &v1.PipelineRun{
+		ObjectMeta: metav1.ObjectMeta{Name: "test-pipeline-run-timeout", Namespace: "foo"},
+	}
+
+	ctx, _ := ttesting.SetupFakeContext(t)
+	cfg := config.NewStore(logtesting.TestLogger(t))
+	ctx = cfg.ToContext(ctx)
+	// Put a span on the context so the timeout helpers create their child spans
+	// against the recording provider.
+	ctx, root := tp.Tracer("test").Start(ctx, "root")
+	defer root.End()
+	c, _ := test.SeedTestData(t, ctx, test.Data{PipelineRuns: []*v1.PipelineRun{pr}})
+
+	if err := timeoutPipelineRun(ctx, logtesting.TestLogger(t), pr, c.Pipeline); err != nil {
+		t.Fatalf("timeoutPipelineRun returned an unexpected error: %v", err)
+	}
+
+	var found bool
+	for _, s := range sr.Ended() {
+		if s.Name() == "timeoutPipelineRun" {
+			found = true
+		}
+	}
+	if !found {
+		t.Errorf("expected a %q span to be recorded, got %v", "timeoutPipelineRun", spanNames(sr.Ended()))
 	}
 }
